+++ stylelint.config.js
@@ -1,0 +1,3 @@
+import config from './src/stylelint.ts'
+
+export default config